+++ app/src/main/res/layout/fragment_write.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.fragment.WriteFragment">
+        tools:context=".view.fragment.DiaryFragment">
 
         <LinearLayout
             android:layout_width="match_parent"
